Fix: Removing a node from the cluster if its stake is unbonded below min threshold (#150)
This PR tries to address a missing gap in the Unbonding process and the
way a node provider can leave a cluster according to the cluster's
terms.

### Problem
At this moment, we keep stakes of active nodes, i.e. Storage and CDN
nodes that are serving a cluster, in `Storages` and `CDNs` maps
accordingly. These maps are supposed to keep stakes that can be slashed
in case of malicious behavior from active nodes.
If an active node wants to chill, i.e. to stop serving the cluster for
some time (ex. for upgrading software or hardware), it executes `chill`
call. Chilling starts according to cluster's terms, i.e. an active node
can be safely stopped only after the `chill_delay` period is passed.
After the `chill_delay`, the node is being removed from the `Storages` /
`CDNs` map and its stake is safe from the slashing.

After the chilling, a node has 2 options: 
1) It can continue to serve the cluster it is assigned to (i.e. execute
`serve` or `store` call) and return back to the `Storages` / `CDNs` map
that contains active nodes.
2) It can start unbonding process for its stake and unlock all tokens
according to cluster's terms, i.e. after the `unbond_delay` period is
passed.

In the current implementation, there is no way to see what nodes decided
to leave the cluster, as the absence of a node in `Storages` / `CDNs`
map doesn't define the exact node status. It can either do chilling and
return back to serving cluster later, or do unbonding and leave the
cluster eventually. In the latter case, the protocol needs to remove the
node from the cluster after the stake is fully withdrawn. Otherwise, we
may have "dead" nodes in the cluster and its topology, that will affect
the routing and other aspects of the cluster.

### Solution
In this PR I introduced 2 additional maps - `LeavingStorages` and
`LeavingCDNs`. These maps contain stakes of nodes that are about to
leave the cluster, i.e. nodes that decided to unbond tokens below the
`min_bond_size` of this cluster.
After the `unbonding_delay` is passed, the node provider can withdraw
all its tokens. It will automatically remove his node from the cluster
based on `LeavingStorages` / `LeavingCDNs` maps.
In the future, `LeavingStorages` / `LeavingCDNs` can be also used to
define specific rules applied to leaving nodes during the unbonding
period.

I tried not to change the existing staking workflow, but extend it with
additional storage items and events. So, the existing workflow should
not be broken.